+++ samples/app/src/main/java/com/example/app/Destinations.kt
@@ -14,6 +14,8 @@ object Destinations {
     const val TOKEN_ROUTE = "token"
     const val MANAGE_WEBAUTHN_KEYS = "webauthnKeys"
     const val MANAGE_USER_KEYS = "userKeys"
+    const val IG = "ig"
+    const val DEVICE_PROFILE = "deviceProfile"
     const val CENTRALIZE_ROUTE = "centralize"
 }